+++ pymia/data/assembler.py
@@ -3,10 +3,8 @@ import pickle
 
 import numpy as np
 
-from . import transformation as tfm
 
-
-class Assembler(metaclass=abc.ABCMeta):
+class Assembler(abc.ABC):
 
     @abc.abstractmethod
     def add_batch(self, prediction, batch: dict):
@@ -17,21 +15,45 @@ def numpy_zeros(shape: tuple, id_: str):
     return np.zeros(shape)
 
 
+def default_sample_fn(params: dict):
+    key = '__prediction'
+    batch = params['batch']
+    idx = params['batch_idx']
+
+    data = params[key]
+    index_expr = batch['index_expr'][idx]
+    if isinstance(index_expr, bytes):
+        # is pickled
+        index_expr = pickle.loads(index_expr)
+
+    return data, index_expr
+
+
 class SubjectAssembler(Assembler):
     """Assembles predictions of one or multiple subjects.
 
     Assumes that the network output, i.e. to_assemble, is of shape (B, ..., C)
-    where B is the batch size and C is the numbers of channels (must be at least 1). ... refers to the arbitrary image
+    where B is the batch size and C is the numbers of channels (must be at least 1) and ... refers to an arbitrary image
     dimension.
     """
 
-    def __init__(self, zero_fn=numpy_zeros):
+    def __init__(self, zero_fn=numpy_zeros, on_sample_fn=default_sample_fn):
+        """Initializes a new instance of the SubjectAssembler class.
+
+        Args:
+            zero_fn: A function that initializes the numpy array to hold the predictions.
+                Args: shape: tuple with the shape of the subject's labels, id_: str identifying the subject.
+                Returns: A np.ndarray
+            on_sample_fn: A function that processes a sample.
+                Args: params: dict with the parameters.
+                Returns tuple of data and index expression, i.e. the processed data and index expression.
+        """
         self.predictions = {}
         self.subjects_ready = set()
         self.zero_fn = zero_fn
+        self.on_sample_fn = on_sample_fn
 
-    def add_batch(self, to_assemble, batch: dict, last_batch=False, transform: tfm.Transform=None):
-        # todo(fabianbalsiger): replace tfm.Transform by callback
+    def add_batch(self, to_assemble, batch: dict, last_batch=False):
         if 'subject_index' not in batch:
             raise ValueError('SubjectAssembler requires "subject_index" to be extracted (use IndexingExtractor)')
         if 'index_expr' not in batch:
@@ -43,7 +65,7 @@ class SubjectAssembler(Assembler):
             to_assemble = {'__prediction': to_assemble}
 
         for idx in range(len(batch['subject_index'])):
-            self.add_sample(to_assemble, batch, idx, transform)
+            self.add_sample(to_assemble, batch, idx)
 
         if last_batch:
             # to prevent from last batch to be ignored
@@ -52,11 +74,11 @@ class SubjectAssembler(Assembler):
     def end(self):
         self.subjects_ready = set(self.predictions.keys())
 
-    def add_sample(self, to_assemble, batch, idx, transform: tfm.Transform=None):
-        # todo(fabianbalsiger): replace tfm.Transform by callback
-        # initialize subject
+    def add_sample(self, to_assemble, batch, idx):
+
         subject_index = batch['subject_index'][idx]
 
+        # initialize subject
         if subject_index not in self.predictions and not self.predictions:
             self.predictions[subject_index] = self._init_new_subject(batch, to_assemble, idx)
         elif subject_index not in self.predictions:
@@ -65,15 +87,12 @@ class SubjectAssembler(Assembler):
 
         for key in to_assemble:
             data = to_assemble[key][idx]
-            if transform is not None:
-                data = transform({key: data, 'batch': batch, 'batch_idx': idx})[key]
+            sample_data, index_expr = self.on_sample_fn({key: data,
+                                                         'batch': batch,
+                                                         'batch_idx': idx,
+                                                         'predictions': self.predictions})
 
-            index_expr = batch['index_expr'][idx]
-            if isinstance(index_expr, bytes):
-                # is pickled
-                index_expr = pickle.loads(index_expr)
-
-            self.predictions[subject_index][key][index_expr.expression] = data
+            self.predictions[subject_index][key][index_expr.expression] = sample_data
 
     def _init_new_subject(self, batch, to_assemble, idx):
         subject_prediction = {}
@@ -103,4 +122,3 @@ class SubjectAssembler(Assembler):
         if '__prediction' in assembled:
             return assembled['__prediction']
         return assembled
-